IL-718 Fix VNET uncommenting in solve script
We need to replace only the first instance of an commented out
`resource_group_name`. This was likely passing tests because until
55adab2c the check for an empty VNET was

 if [ $VNET != "azurerm_virtual_network" ]

so if $VNET was actually an empty string, to bash it looked like

 if [ != "azurerm_virtual_network" ]

so bash would error with "bash: [: !=: unary operator expected"
but our script would grind along and "exit 0"

Now that we check properly, we catch that we were uncommenting
multiple lines of `resource_group_name`, which was invalid
Terraform code, so the VNET wasn't actually created.